+++ fugue/src/main/java/io/atlassian/fugue/Effect.java
@@ -19,22 +19,23 @@ import java.util.function.Consumer;
 
 /**
  * A thing that performs a side-effect.
- * 
+ *
  * @param <A> the type that this Effect takes as input
- * 
+ *
  * @since 1.0
+ * @deprecated in favour of {@link Consumer}
  */
-@FunctionalInterface public interface Effect<A> extends Consumer<A> {
+@Deprecated @FunctionalInterface public interface Effect<A> extends Consumer<A> {
   /**
    * Perform the side-effect.
-   * 
+   *
    * @param a the input to use for performing the effect.
    */
   void apply(A a);
 
   /**
    * Adapt to the Java 8 interface.
-   * 
+   *
    * @param a the input to use for performing the effect.
    * @since 3.0
    */
@@ -44,15 +45,20 @@ import java.util.function.Consumer;
 
   /**
    * A thing upon which side-effects may be applied.
-   * 
+   *
    * @param <A> the type of thing to supply to the effect.
+   *
+   * @deprecated in favour of {@link Iterable}
    */
-  @FunctionalInterface interface Applicant<A> {
+  @Deprecated @FunctionalInterface interface Applicant<A> {
     /**
      * Perform the given side-effect for each contained element.
-     * 
+     *
      * @param effect the input to use for performing the effect.
+     *
+     * @deprecated extend or implement {@link Iterable#forEach(Consumer)}
+     * instead
      */
-    void foreach(Effect<? super A> effect);
+    @Deprecated void foreach(Effect<? super A> effect);
   }
 }