+++ fugue/src/test/java/io/atlassian/fugue/EitherRightBiasTest.java
@@ -119,18 +119,20 @@ public class EitherRightBiasTest {
 
   @Test public void foreachRight() {
     final AtomicBoolean called = new AtomicBoolean(false);
+    @SuppressWarnings("deprecation")
     final Effect<Integer> effect = integer -> called.set(true);
 
-    r.foreach(effect);
+    r.forEach(effect);
 
     assertThat(called.get(), is(true));
   }
 
   @Test public void foreachLeft() {
     final AtomicBoolean called = new AtomicBoolean(false);
+    @SuppressWarnings("deprecation")
     final Effect<Integer> effect = integer -> called.set(true);
 
-    l.foreach(effect);
+    l.forEach(effect);
 
     assertThat(called.get(), is(false));
   }